+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityBreezeWindChargePhysicsComponentImpl.java
@@ -2,6 +2,7 @@ package org.allaymc.server.entity.component.projectile;
 
 import org.allaymc.api.world.sound.CustomSound;
 import org.allaymc.api.world.sound.SoundNames;
+import org.allaymc.api.world.particle.CustomParticle;
 
 /**
  * Breeze wind charge projectile behavior.
@@ -22,5 +23,10 @@ public class EntityBreezeWindChargePhysicsComponentImpl extends EntityWindCharge
     @Override
     protected CustomSound getBurstSound() {
         return new CustomSound(SoundNames.BREEZE_WIND_CHARGE_BURST);
+    }
+
+    @Override
+    protected CustomParticle getBurstParticle() {
+        return new CustomParticle("minecraft:breeze_wind_explosion");
     }
 }